Cambio 1 en vista solicitudID y cambio 8 de vista incremento
