+++ src/app/services/api/api.service.ts
@@ -49,7 +49,7 @@ export class ApiService {
       },
     };
 
-    // Retornamos los datos como un Observable simulando la respuesta de un API
+    //Retornamos los datos como un Observable simulando la respuesta de un API
     return of(response);
   }
 
@@ -167,12 +167,12 @@ export class ApiService {
           label: '1',
           active: true,
         },
-        
+
       ],
     };
 
     return of(response);
   }
-  
-  
+
+
 }
